backend/requirements.txt
fastapi
uvicorn
fastmcp
langchain
langchain-core
langchain-google-genai
langchain-openai
langchain-classic
langchain-mcp-adapters
mesa
networkx
numpy
python-dotenv
